fix: remove lazy loading of array and blocks
